% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.ext.R
\name{add_ass_names}
\alias{add_ass_names}
\title{Add the names of assemblies}
\usage{
add_ass_names(Fcal, Fprd, Fobs, assMotifs, AssNames)
}
\arguments{
\item{Fcal}{gghhh}

\item{Fprd}{gghhh}

\item{Fobs}{gghhh}

\item{assMotifs}{gghhh}

\item{AssNames}{gghhh}
}
\value{
gghhh
}
\description{
llll
}
\details{
dd
}
\keyword{internal}
