% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.int.R
\name{affect_motifs}
\alias{affect_motifs}
\title{Cluster the assemblages by assembly motifs}
\usage{
affect_motifs(affectElt, mOccur)
}
\arguments{
\item{affectElt}{gg}

\item{mOccur}{gg}
}
\value{
ff
}
\description{
fff
}
\details{
dd
}
\keyword{internal}
