% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{AIC}
\alias{AIC}
\title{Compute the AIC from the RSS value.}
\usage{
AIC(x, y, nbK)
}
\arguments{
\item{x}{A logical/numeric vector.}

\item{y}{A logical/numeric vector.}

\item{nbK}{A positive scalar.}
}
\value{
Return the the Akaike Information Criterion (AIC).
}
\description{
`AIC` returns the Akaike Information Criterion (AIC) computed from the Residuals Sum of Square and the number of used parameters.
}
\details{
None.
}
