% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{AICc}
\alias{AICc}
\title{AICc of two numeric vectors}
\usage{
AICc(x, y, nbK)
}
\arguments{
\item{x, y}{numeric vectors of same length.}

\item{nbK}{an integer. \code{nbK}should be higher than \code{length(x)}: return an error if not.}
}
\value{
Return the the Akaike Information Criterion (AIC).
}
\description{
Take two numeric vectors and return the Akaike Information
 Criterion (AICc) computed from the Residuals Sum of Square and
  the number of used parameters, and corrected for .
}
\details{
Be careful, the function \code{pmse} is not symmetrical.
The first argument is vector of estimated, modelled or predicted values,
 the second argument is the vector of reference.
}
\keyword{internal}
