% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{amean}
\alias{amean}
\title{Arithmetic mean}
\usage{
amean(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the arithmetic mean of a vector, as standard 'mean' function.
}
\description{
Take a numeric vector and return the arithmetic mean.
}
\details{
`amean` function does as the standard `stats::mean` function, but with option `na.rm = TRUE` by default.
}
\keyword{internal}
