% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{amean_byelt_LOO}
\alias{amean_byelt_LOO}
\title{Arithmetic mean (amean) by elements occurring within
the assemblage and other assemblages belonging to a same motif (byelt),
except the assemblage to predict (leave-one-out, LOO)}
\usage{
amean_byelt_LOO(fctMot, mOccurMot)
}
\arguments{
\item{fctMot}{a vector of numeric values (assembly properties).}

\item{mOccurMot}{a matrix of occurrence (occurrence of elements).
Its first dimension equals to \code{length(fctMot)}. Its second
dimension equals to the number of elements.}
}
\value{
Return a vector of \code{length(fct)}, of which values are
 computed as an arithmetic mean of all vector elements over several
  experiments, excepted the value of element to predict that have
  been left out.
}
\description{
The numeric vector \code{fctMot} get together the
 properties of assemblages belonging to a same assembly motif.
 The properties \code{fctMot} of assemblages containing a given element
  are separately averaged, except the assemblage to predict. The property
   of each assemblage is computed as the average of mean values of
   assemblages containing the same elements as the considered assemblage.
    The elemental composition of each assemblage is specified in the
     binary matrix \code{mOccurMot}: \code{0} if the element does not
      occur, \code{1} if the element occurs.
}
\details{
Prediction is computed using arithmetic mean \code{amean}
by element \code{byelt} occurring within the assemblage and other
 assemblages of a same motif, by including all assemblages belonging
 to a same motif, except the assemblage to predict.
}
\keyword{internal}
