% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{amean_bymot_LOO}
\alias{amean_bymot_LOO}
\title{Arithmetic mean (amean) by motif (bymot) by leave-one-out (LOO)}
\usage{
amean_bymot_LOO(fctMot)
}
\arguments{
\item{fctMot}{a vector of numeric values of elements belonging to a same
motif.}
}
\value{
Return a vector of same length as \code{fctMot}, of which values
are computed as the arithmetic mean of all vector elements, excepted the
value of element to predict that have been left out.
}
\description{
Take the numeric vector \code{fctMot} and return a vector of
same length, of which values are computed as the arithmetic mean of all
vector elements.
}
\details{
Prediction computed using arithmetic mean \code{amean} by motif
 \code{bymot} in a whole (WITHOUT taking into account species
 contribution) by excluding the element to predict.
}
\keyword{internal}
