% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{argmax}
\alias{argmax}
\title{Index of the maximum values of a vector}
\usage{
argmax(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the index of the maximum values of a vector.
}
\description{
Take a numeric vector and return the index of the
 maximum values.
}
\details{
The function \code{argmax} works as the standard function
 \code{which.max}.
}
\keyword{internal}
