% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{argmin}
\alias{argmin}
\title{Returns the index of the minimum values of a vector.}
\usage{
argmin(x)
}
\arguments{
\item{x}{A logical/numeric vector.}
}
\value{
Return the index of the minimum values of a vector.
}
\description{
`argmin` returns the index of the minimum values of a vector.
}
\details{
`argmin` function is as the standard `which.min` function.
}
