% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{asd}
\alias{asd}
\title{Arithmetical standard deviation}
\usage{
asd(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the arithmetic standard deviation of a vector,
 as the standard function \code{stats::sd}.
}
\description{
Take a numeric vector and return the arithmetic standard
 deviation.
}
\details{
The fucntion \code{asd} works as the standard function \code{sd},
 but with option \code{na.rm = TRUE} by default.
}
\keyword{internal}
