% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.ext.R
\name{cluster_elements}
\alias{cluster_elements}
\title{fffff}
\usage{
cluster_elements(mOccur, fct,
                        xpr = rep(1, length(fct)),
                        opt.meth = "divisive",
                        opt.mean = "amean",
                        opt.mod  = "byelt",
                        opt.prn  = TRUE)
}
\arguments{
\item{mOccur}{ff}

\item{fct}{ff}

\item{xpr}{ff}

\item{opt.meth}{gg}

\item{opt.mean}{gg}

\item{opt.mod}{gg}

\item{opt.prn}{gg}
}
\value{
gg
}
\description{
ddd
}
