% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.ext.R
\name{cluster_elements_LOO}
\alias{cluster_elements_LOO}
\title{rrr}
\usage{
cluster_elements_LOO(mOccur, fct,
                            xpr      = rep(1, length(fct)),
                            opt.meth = "divisive",
                            opt.mean = "amean",
                            opt.mod  = "byelt",
                            opt.prn  = TRUE)
}
\arguments{
\item{mOccur}{gg}

\item{fct}{gf}

\item{xpr}{gg}

\item{opt.meth}{gg}

\item{opt.mean}{gg}

\item{opt.mod}{gg}

\item{opt.prn}{gg}
}
\value{
gg
}
\description{
ddd
}
