% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.ext.R
\name{cluster_motifs}
\alias{cluster_motifs}
\title{kkk}
\usage{
cluster_motifs(affectElt, mOccur, fct,
                      opt.meth   = "divisive",
                      nbclMotMax = "NA",
                      opt.prn    = FALSE       )
}
\arguments{
\item{affectElt}{gg}

\item{mOccur}{gg}

\item{fct}{gg}

\item{opt.meth}{ff}

\item{nbclMotMax}{gg}

\item{opt.prn}{gg}
}
\value{
gg
}
\description{
fff
}
