% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.int.R
\name{compact_index}
\alias{compact_index}
\title{Return a vector continuously indexed from 1 to max(index)}
\usage{
compact_index(v)
}
\arguments{
\item{v}{gg}
}
\value{
gg
}
\description{
ddd
}
\details{
dd
}
\keyword{internal}
