% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.ext.R
\name{compute_fit_stats}
\alias{compute_fit_stats}
\title{Compute the matrix 3 x nbclusters of R2cal, R2prd, and missing values}
\usage{
compute_fit_stats(mCal, mPrd, fct, nbK)
}
\arguments{
\item{mCal}{gghhh}

\item{mPrd}{gghhh}

\item{fct}{gghhh}

\item{nbK}{gghhh}
}
\value{
gghhh
}
\description{
Take a numeric f.
}
\details{
dd
}
\keyword{internal}
