% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.ext.R
\name{compute_motif_stats}
\alias{compute_motif_stats}
\title{Compute the statistiques of each motifs}
\usage{
compute_motif_stats(tCal, tPrd, mMotifs)
}
\arguments{
\item{tCal}{gghhh}

\item{tPrd}{gghhh}

\item{mMotifs}{gghhh}
}
\value{
gghhh
}
\description{
Take a numeric f.
}
\details{
dd
}
\keyword{internal}
