% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.ext.R
\name{compute_tree_stats}
\alias{compute_tree_stats}
\title{Compute and plot the best Prediction
                       as a concatenation of the best predictions}
\usage{
compute_tree_stats(mCal, mPrd, mStats, fct, nbK)
}
\arguments{
\item{mCal}{gghhh}

\item{mPrd}{gghhh}

\item{mStats}{gghhh}

\item{fct}{gghhh}

\item{nbK}{gghhh}
}
\description{
Take a numeric f.
}
\details{
dd
}
\keyword{internal}
