% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{confidence.intervals}
\alias{confidence.intervals}
\title{Confidence and prediction intervals for a linear model}
\usage{
confidence.intervals(model, pvalue = 0.05,
                            xlim = range(model$model[ , 2]), x.length = 20)
}
\arguments{
\item{model}{a linear model obtained using `lm` function.}

\item{pvalue}{a p-value. `pvalue = 0.05` by default.}

\item{xlim}{a vector of two numeric values. `xlim = range(model$model[,2])` by default.}

\item{x.length}{an integer, indicating the number of points to compute.}
}
\value{
Return a `list` containing a vector `x` of x-values (ranging from `xlim[1]` to `xlim[2]`), the corresponding values `y.confidence` for confidence interval, the corresponding values `y.prediction` for prediction interval.
}
\description{
Compute the confidence and prediction intervals for a linear regression model `model`.
}
\details{
None.
}
