% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{confidence_intervals}
\alias{confidence_intervals}
\title{Confidence and prediction intervals for a linear model}
\usage{
confidence_intervals(model, pvalue = 0.05,
                            xlim = range(model$model[ , 2]), x.length = 20)
}
\arguments{
\item{model}{a linear model obtained using \code{lm} function.}

\item{pvalue}{a p-value. \code{pvalue = 0.05} by default.}

\item{xlim}{a vector of two numeric values.
\code{xlim = range(model$model[,2])} by default.}

\item{x.length}{an integer, indicating the number of points to compute.}
}
\value{
Return a \code{list} containing a vector \code{x} of x-values
(ranging from \code{xlim[1]} to \code{xlim[2]}), the corresponding values
 \code{y.confidence} for confidence interval, the corresponding values
  \code{y.prediction} for prediction interval.
}
\description{
Compute the confidence and prediction intervals
for a linear regression model \code{model}.
}
\details{
None.
}
\keyword{internal}
