% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{cor2}
\alias{cor2}
\title{Return the Pearson' R2 of a linear regression}
\usage{
cor2(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{ff}

\item{y}{fff}

\item{na.rm}{ff}
}
\value{
ff
}
\description{
ddd
}
\keyword{internal}
