% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{first_argmax}
\alias{first_argmax}
\title{Index of the first maximum value of a vector}
\usage{
first_argmax(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the index of the first (the lowest index)
 maximum values of a vector.
}
\description{
Take a numeric vector and return the index of
the first (the lowest index) maximum values.
}
\details{
The function \code{first_argmax} works as the
standard function \code{min(which.max)}.
}
\keyword{internal}
