% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{gmean}
\alias{gmean}
\title{Geometric mean}
\usage{
gmean(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the geometric mean of the vector \code{x}.
}
\description{
Take a numeric vector and return its geometric mean.
}
\details{
None.
}
\keyword{internal}
