% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{gmean_byelt}
\alias{gmean_byelt}
\title{Geometric mean (gmean) by elements occurring within the assemblage
and other assemblages, all assemblages belonging to a same motif (byelt)}
\usage{
gmean_byelt(fctMot, mOccurMot)
}
\arguments{
\item{fctMot}{a vector of numeric values (assembly properties).}

\item{mOccurMot}{a matrix of occurrence (occurrence of elements).
Its first dimension equals to \code{length(fctMot)}. Its second
 dimension equals to the number of elements.}
}
\value{
Return a vector of \code{length(fctMot)}, of which values
 are computed as an geometric mean.
}
\description{
The numeric vector \code{fctMot} get together the properties
 of assemblages belonging to a same assembly motif. The properties
 \code{fctMot} of assemblages containing a given element are separately
  averaged. The property of each assemblage is computed as the average
   of mean values of assemblages containing the same elements as the
    considered assemblage. The elemental composition of each assemblage
     is specified in the binary matrix \code{mOccurMot}: \code{0}
      if the element does not occur, \code{1} if the element occurs.
}
\details{
Prediction is computed using geometric mean \code{gmean}
 by element \code{byelt} occurring within the assemblage and other
  assemblages of a same motif, by including all assemblages belonging
  to a same motif, even the one to predict.
}
\keyword{internal}
