% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{gmean_bymot_LOO}
\alias{gmean_bymot_LOO}
\title{Geometric mean (gmean) by motif (bymot) by leave-one-out (LOO)}
\usage{
gmean_bymot_LOO(fctMot)
}
\arguments{
\item{fctMot}{a vector of numeric values of elements belonging
to a same motif.}
}
\value{
Return a vector of \code{length(fctMot)}, of which values
are computed as the geometric mean of all vector elements, excepted
 the value of element to predict that have been left out.
}
\description{
Take the numeric vector \code{fctMot} and return a vector
 of same length, of which values are computed as the geometric mean
 of all vector elements.
}
\details{
Prediction computed using geometric mean \code{gmean}
by motif \code{bymot} in a whole (WITHOUT taking into account
 species contribution) by excluding the element to predict.
}
\keyword{internal}
