% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{gsd}
\alias{gsd}
\title{Geometric standard deviation}
\usage{
gsd(x)
}
\arguments{
\item{x}{A logical/numeric vector.}
}
\value{
Return the geometric standard deviation of a vector.
}
\description{
`gsd` takes a logical/numeric vector and returns its geometric  standard deviation.
}
\details{
None.
}
