% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.int.R
\name{maffect_motifs}
\alias{maffect_motifs}
\title{Cluster the assemblages by assembly motifs}
\usage{
maffect_motifs(tree, mOccur)
}
\arguments{
\item{tree}{ff}

\item{mOccur}{ff}
}
\value{
gg
}
\description{
fff
}
\details{
dd
}
\keyword{internal}
