% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{mean_fct}
\alias{mean_fct}
\title{Switch for arithmetic or geometric mean according to option "opt.mean"}
\usage{
mean_fct(x, opt.mean = c("amean", "gmean"))
}
\arguments{
\item{x}{a numeric vector.}

\item{opt.mean}{equals `amean` or `gmean` according to that mean value must be computed with arithmetic or geometric formula. There is no default value.}
}
\value{
Return the arithmetic or geometric mean of the vector `x` according to opt.mean.
}
\description{
Take the numeric vector and return its weighted geometric standard deviation.
}
\details{
None.
}
