% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{mse}
\alias{mse}
\title{Compute the Mean Square Error (MSE) of two numeric vectors}
\usage{
mse(x, y)
}
\arguments{
\item{x}{A logical/numeric vector.}

\item{y}{A logical/numeric vector.}
}
\value{
Return the Mean Square Error (MSE).
}
\description{
`mse` returns the Mean Square Error (MSE) of two numeric vectors.
}
\details{
None.
}
