% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separating.ext.R
\name{multiplicative_decomposition}
\alias{multiplicative_decomposition}
\title{Decompose the observed performance into a.inter and a.comp ratios}
\usage{
multiplicative_decomposition(mOccur, Fobs, rm.mono = TRUE)
}
\arguments{
\item{mOccur}{ff}

\item{Fobs}{dd}

\item{rm.mono}{dd}
}
\value{
dd
}
\description{
bbb
}
