% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.int.R
\name{names_assembly}
\alias{names_assembly}
\title{names (in minus letters) of elements,}
\usage{
names_assembly(affectElt, mOccur)
}
\arguments{
\item{affectElt}{tt}

\item{mOccur}{ttt}
}
\value{
tt
}
\description{
bbb
}
\details{
dd
}
\keyword{internal}
