% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.ext.R
\name{plot_prediction}
\alias{plot_prediction}
\title{Plot all the necessary predicting figures}
\usage{
plot_prediction(res,
                        xpr.who  = 5,
                        nbOpt    = 0,
                        titre    = "",
                        clu.stat = FALSE,
                        clu.cal  = FALSE,
                        clu.prd  = FALSE,
                        tre.stat = FALSE,
                        tre.prd  = FALSE,
                        tre.pub  = FALSE,
                        tre.best = FALSE,
                        tre.opt  = TRUE,
                        tre.calvsprd = FALSE,
                        ass.loc  = FALSE,
                        opt.aov  = FALSE, pvalue = 0.025,
                        opt.all  = FALSE  )
}
\arguments{
\item{res}{gg}

\item{xpr.who}{gg}

\item{nbOpt}{gg}

\item{titre}{gg}

\item{clu.stat}{gg}

\item{clu.cal}{gg}

\item{clu.prd}{gg}

\item{tre.stat}{gghhh}

\item{tre.prd}{gghhh}

\item{tre.pub}{gghhh}

\item{tre.best}{gghhh}

\item{tre.opt}{gghhh}

\item{tre.calvsprd}{gghhh}

\item{ass.loc}{gghhh}

\item{opt.aov}{gghhh}

\item{pvalue}{gghhh}

\item{opt.all}{gghhh}
}
\value{
gghhh
}
\description{
ddd
}
\details{
dd
}
