% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.ext.R
\name{plot_prediction_LOO}
\alias{plot_prediction_LOO}
\title{mmm}
\usage{
plot_prediction_LOO(Fcal, Fprd, Fobs, assMotifs, nbcl,
                           xylim    = range(Fobs),
                           titre    = "",
                           opt.mean = "amean",
                           opt.aov  = FALSE,  pvalue   = 0.05)
}
\arguments{
\item{Fcal}{gghhh}

\item{Fprd}{gghhh}

\item{Fobs}{gghhh}

\item{assMotifs}{gghhh}

\item{nbcl}{gghhh}

\item{xylim}{gghhh}

\item{titre}{gghhh}

\item{opt.mean}{gghhh}

\item{opt.aov}{gghhh}

\item{pvalue}{gghhh}
}
\value{
gghhh
}
\description{
gghhh
}
\details{
dd
}
