% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.ext.R
\name{plot_prediction_simple}
\alias{plot_prediction_simple}
\title{Plot Predictions vs Observations}
\usage{
plot_prediction_simple(Fprd, Fobs, assMotifs, nbcl,
                              titre    = "",
                              xylim    = range(Fobs),
                              opt.mean = "amean",
                              opt.aov  = FALSE,  pvalue   = 0.025)
}
\arguments{
\item{Fprd}{gghhh}

\item{Fobs}{gghhh}

\item{assMotifs}{gghhh}

\item{nbcl}{gghhh}

\item{titre}{gghhh}

\item{xylim}{gghhh}

\item{opt.mean}{gghhh}

\item{opt.aov}{gghhh}

\item{pvalue}{gghhh}
}
\value{
gghhh
}
\description{
gggg
}
\details{
dd
}
