% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.ext.R
\name{plot_prediction_supp}
\alias{plot_prediction_supp}
\title{Plot all the necessary predicting figures}
\usage{
plot_prediction_supp(res, nbOpt,
                            titre    = "",
                            sup.prd  = FALSE,
                            sup.pub  = FALSE,
                            sup.opt  = TRUE,
                            ass.loc  = FALSE,
                            opt.all  = FALSE  )
}
\arguments{
\item{res}{gg}

\item{nbOpt}{gg}

\item{titre}{gg}

\item{sup.prd}{gg}

\item{sup.pub}{gg}

\item{sup.opt}{gg}

\item{ass.loc}{gg}

\item{opt.all}{gg}
}
\value{
gg
}
\description{
ddd
}
\details{
dd
}
