% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.ext.R
\name{plot_stats}
\alias{plot_stats}
\title{Plot the matrix 3 x nbclusters of R2cal, R2prd, and missing values}
\usage{
plot_stats(mStats, nbElt, titre = "")
}
\arguments{
\item{mStats}{gghhh}

\item{nbElt}{gghhh}

\item{titre}{gghhh}
}
\value{
gghhh
}
\description{
Take a numeric f.
}
\details{
dd
}
