% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.int.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot a Divisive or Agglomerative Hierarchical Clustering}
\usage{
plot_tree(tree, signifElt = NULL, col = "black", titre = "")
}
\arguments{
\item{tree}{kk}

\item{signifElt}{ll}

\item{col}{ll}

\item{titre}{ll}
}
\value{
ll
}
\description{
ddd
}
\details{
dd
}
\keyword{internal}
