% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{pmse}
\alias{pmse}
\title{Probability associated with the Coefficient of Determination (R2)
 of two numeric vectors}
\usage{
pmse(prd, obs, nbK)
}
\arguments{
\item{prd}{a numeric vector of predicted values.}

\item{obs}{a numeric vector of reference values of \code{length(prd)}.}

\item{nbK}{an integer. \code{nbK}should be higher than 1: return an
error if not.}
}
\value{
Return the Probability associated with the Coefficient
of Determination (R2).
}
\description{
Take two numeric vectors and return the
 Probability associated with their Coefficient of Determination (R2).
}
\details{
Be careful, the function \code{pmse} is not symmetrical.
The first argument \code{prd} is vector of estimated, modelled
or predicted values, the second argument \code{obs} is
the vector of reference.
}
\keyword{internal}
