% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{predict_amean_byelt_supp}
\alias{predict_amean_byelt_supp}
\title{Prediction of supplementary assemblages computed}
\usage{
predict_amean_byelt_supp(appFct, appMotifs, appOccur,
                             supMotifs, supOccur )
}
\arguments{
\item{appFct}{cccc}

\item{appMotifs}{cccc}

\item{appOccur}{cccc}

\item{supMotifs}{cccc}

\item{supOccur}{cccc}
}
\value{
cccc
}
\description{
Take a numeric f.
}
\details{
dd
}
\keyword{internal}
