% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{predict_amean_bymot}
\alias{predict_amean_bymot}
\title{Arithmetic mean (amean) by assembly motif (bymot)}
\usage{
predict_amean_bymot(fct, assMotifs)
}
\arguments{
\item{fct}{a vector of numeric values (assembly properties).}

\item{assMotifs}{a vector of labels of same length as \code{fct} (assembly
motifs).}
}
\value{
Return a vector of same length as \code{fct}, of which values are
 computed as the arithmetic mean of all vector elements belonging to a same
 motif.
}
\description{
Take the numeric vector \code{fct} and return a vector of same
 length, of which values are computed as the arithmetic mean of all vector
  elements belonging to a same motif. The motif of each vector element is
   specified in the vector \code{assMotifs}.
}
\details{
Prediction is computed using arithmetic mean \code{amean} by motif
 \code{bymot} in a whole (WITHOUT taking into account species contribution)
  by including all elements belonging to a same motif, even the one to
  predict.
}
\keyword{internal}
