% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{predict_amean_bymot_jack_xpr}
\alias{predict_amean_bymot_jack_xpr}
\title{Arithmetic mean (amean) by motif (bymot) by jackknife (jack)
over several experiments (xpr)}
\usage{
predict_amean_bymot_jack_xpr(fct, assMotifs, jack, xpr)
}
\arguments{
\item{fct}{a vector of numeric values (assembly properties).}

\item{assMotifs}{a vector of labels of same length as \code{fct}
(assembly motifs).}

\item{jack}{a vector of two elements. The first one \code{jack[1]}
specifies the size of subset, the second one \code{jack[2]} specifies
 the number of subsets.}

\item{xpr}{a vector of labels of same length as \code{fct}
(assembly experiments).}
}
\value{
Return the arithmetic mean of a vector, as standard \code{mean}
 function.
}
\description{
Take a numeric vector and return the predicted vector
 computed as the arithmetic mean of all elements belonging to a same motif.
}
\details{
Prediction is computed using arithmetic mean \code{amean}
by motif \code{bymot} in a whole (WITHOUT taking into account species
 contribution). The elements belonging to a same motif are divided
 into \code{jack[2]} subsets of \code{jack[1]} elements. Prediction
  is computed by excluding \code{jack[1]} elements, of which the
  element to predict. If the total number of elements belonging to
  the motif is lower than \code{jack[1]*jack[2]}, prediction is
  computed by Leave-One-Out (LOO).
}
\keyword{internal}
