% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{predict_amean_bymot_LOO_xpr}
\alias{predict_amean_bymot_LOO_xpr}
\title{Arithmetic mean (amean) by motif (bymot) by leave-one-out (LOO)
 over several experiments (xpr)}
\usage{
predict_amean_bymot_LOO_xpr(fct, assMotifs, xpr)
}
\arguments{
\item{fct}{a vector of numeric values (assembly properties).}

\item{assMotifs}{a vector of labels of same length as \code{fct}
(assembly motifs).}

\item{xpr}{a vector of labels of same length as \code{fct}
(assembly experiments).}
}
\value{
Return a vector of same length as \code{fctMot}, of which
values are computed as the arithmetic mean of all vector elements,
 over  several experiments, excepted the value of element to predict
 that have been left out.
}
\description{
Take a numeric vector and return the predicted vector
computed as the arithmetic mean of all elements belonging to a same motif.
}
\details{
Prediction computed using arithmetic mean \code{amean} by
motif \code{bymot} in a whole (WITHOUT taking into account species
 contribution) by excluding the element to predict.
}
\keyword{internal}
