% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{predict_amean_bymot_supp}
\alias{predict_amean_bymot_supp}
\title{Prediction of supplementary assemblages}
\usage{
predict_amean_bymot_supp(appFct, appMotifs, supMotifs)
}
\arguments{
\item{appFct}{a vector of numeric values (assembly properties).}

\item{appMotifs}{a vector of labels of \code{length(appFct)}
(assembly motifs).}

\item{supMotifs}{a vector of labels of assembly motifs of which values must be predicted.}
}
\value{
Return a vector of \code{length(supMotifs)}. The values are
 computed using arithmetic mean of elements belonging to \code{appMotifs}
 and sharing a same motif.
}
\description{
Take a numeric f.
}
\details{
Prediction ...
}
\keyword{internal}
