% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{predict_amean_bymot_xpr}
\alias{predict_amean_bymot_xpr}
\title{Arithmetic mean (amean) by motif (bymot) over several experiments
(xpr)}
\usage{
predict_amean_bymot_xpr(fct, assMotifs, xpr)
}
\arguments{
\item{fct}{a vector of numeric values (assembly properties).}

\item{assMotifs}{a vector of labels of same length as \code{fct}
(assembly motifs).}

\item{xpr}{a vector of labels of same length as \code{fct}
(assembly experiments).}
}
\value{
Return a vector of same length as \code{fct}, of which values
are computed as the arithmetic mean of all vector elements belonging to
 the same motif over several experiments.
}
\description{
Take the numeric vector \code{fct} and return a vector of same
length, of which values are computed as the arithmetic mean of all vector
elements belonging to a same motif, over several experiments. The motif of
 each vector element is specified in the vector \code{assMotifs}. The
  experiment of each vector element is specified in the vector \code{xpr}.
}
\details{
Prediction is computed using arithmetic mean \code{amean} by
motif \code{bymot} in a whole (WITHOUT taking into account species
contribution) by including all elements belonging to a same motif,
 even the one to predict.
}
\keyword{internal}
