% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.ext.R
\name{predict_function}
\alias{predict_function}
\title{Compute assembly functioning and associated statistiques
         using the Clustering model}
\usage{
predict_function(tree.cal, mOccur, fct,
                        xpr = rep(1, length(fct)),
                        opt.mean = "amean",
                        opt.mod  = "byelt",
                        opt.jack = FALSE,  jack = c(2,5))
}
\arguments{
\item{tree.cal}{gghhh}

\item{mOccur}{gghhh}

\item{fct}{gghhh}

\item{xpr}{gghhh}

\item{opt.mean}{gghhh}

\item{opt.mod}{gghhh}

\item{opt.jack}{gghhh}

\item{jack}{gghhh}
}
\value{
gghhh
}
\description{
Take a numeric f.
}
\details{
dd
}
