% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.ext.R
\name{predict_function_supp}
\alias{predict_function_supp}
\title{#  Compute assembly functioning and associated statistiques
         using the Clustering model
         for Supplementary Assemblies by knowing their elemental Composition}
\usage{
predict_function_supp(tree, res.prd,
                             supOccur, appOccur, appFct,
                             opt.mean = "amean",
                             opt.mod  = "byelt"    )
}
\arguments{
\item{tree}{gghhh}

\item{res.prd}{gghhh}

\item{supOccur}{gghhh}

\item{appOccur}{gghhh}

\item{appFct}{gghhh}

\item{opt.mean}{gghhh}

\item{opt.mod}{gghhh}
}
\value{
gghhh
}
\description{
Take a numeric f.
}
\details{
dd
}
