% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{predict_gmean_bymot}
\alias{predict_gmean_bymot}
\title{Geometric mean (gmean) by motif (bymot)}
\usage{
predict_gmean_bymot(fct, assMotifs)
}
\arguments{
\item{fct}{a vector of numeric values (assembly properties).}

\item{assMotifs}{a vector of labels of \code{length(fct)}
(assembly motifs).}
}
\value{
Return a vector of \code{length(fct)}, of which values are
computed as the geometric mean of all vector elements belonging
to a same motif.
}
\description{
Take the numeric vector \code{fct} and return a vector
of same length, of which values are computed as the geometric mean
of all vector elements belonging to a same motif. The motif of each
vector element is specified in the vector \code{assMotifs}.
}
\details{
Prediction is computed using geometric mean \code{gmean}
by motif \code{bymot} in a whole (WITHOUT taking into account
species contribution) by including all elements belonging
to a same motif, even the one to predict.
}
\keyword{internal}
