% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{predict_gmean_bymot_jack}
\alias{predict_gmean_bymot_jack}
\title{Geometric mean (gmean) by motif (bymot) by jackknife (jack)}
\usage{
predict_gmean_bymot_jack(fct, assMotifs, jack)
}
\arguments{
\item{fct}{a vector of numeric values (assembly properties).}

\item{assMotifs}{a vector of labels of same length as \code{fct}
(assembly motifs).}

\item{jack}{a vector of two elements. The first one\code{jack[1]}
specifies the size of subset, the second one \code{jack[2]} specifies
 the number of subsets.}
}
\value{
Return a vector of same length as \code{fct}, of which
values are computed as the geometric mean of all vector elements.
}
\description{
Take a numeric vector and return the predicted vector
 computed as the geometric mean of all elements belonging to the same motif.
}
\details{
Prediction is computed using geometric mean \code{gmean}
by motif \code{bymot} in a whole (WITHOUT taking into account species
 contribution). The elements belonging to a same motif are divided
 into \code{jack[2]} subsets of \code{jack[1]} elements. Prediction
  is computed by excluding \code{jack[1]} elements, of which the
  element to predict. If the total number of elements belonging
  to the motif is lower than \code{jack[1]*jack[2]}, prediction
   is computed by Leave-One-Out (LOO).
}
\keyword{internal}
