% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{predict_gmean_bymot_LOO_xpr}
\alias{predict_gmean_bymot_LOO_xpr}
\title{Geometric mean (gmean) by motif (bymot) by leave-one-out (LOO)
 over several experiments (xpr)}
\usage{
predict_gmean_bymot_LOO_xpr(fct, assMotifs, xpr)
}
\arguments{
\item{fct}{a vector of numeric values (assembly properties).}

\item{assMotifs}{a vector of labels of same length as \code{fct}
(assembly motifs).}

\item{xpr}{a vector of labels of same length as \code{fct}
(assembly experiments).}
}
\value{
Return a vector of same length as \code{fct},
of which values are computed as the geometric mean of all
vector elements, over several experiments, excepted the value
 of element to predict that have been left out.
}
\description{
Take a numeric vector and return the predicted vector
computed as the gGeometric mean of all elements belonging to a same motif.
}
\details{
Prediction computed using geometric mean \code{gmean}
 by motif \code{bymot} in a whole (WITHOUT taking into
 account species contribution) by excluding the element to predict.
}
\keyword{internal}
