% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.int.R
\name{predict_supp}
\alias{predict_supp}
\title{Arithmetic mean (amean) by motif (bymot) by jackknife (jack) over
 several experiments (xpr)}
\usage{
predict_supp(appFct, appMotifs, appOccur,
                           supMotifs, supOccur,
                           opt.mean = "amean",
                           opt.mod  = "bymot"  )
}
\arguments{
\item{appFct}{a vector of numeric values (assembly properties).}

\item{appMotifs}{a vector of labels of \code{length(fct)} (assembly motifs).}

\item{appOccur}{a matrix of occurrence (occurrence of elements).
Its first dimension equals to \code{length(fct)}. Its second dimension
 equals to the number of elements.}

\item{supMotifs}{a vector of labels of \code{length(fct)} (assembly motifs).}

\item{supOccur}{a matrix of occurrence (occurrence of elements).
Its first dimension equals to \code{length(fct)}. Its second dimension
 equals to the number of elements.}

\item{opt.mean}{equal to \code{"amean"} (by default) or \code{"gmean"}.}

\item{opt.mod}{equal to \code{"bymot"} (by default) or \code{"byelt"}.}
}
\value{
Return the arithmetic mean of a vector, as standard \code{mean}
 function.
}
\description{
Take a numeric vector and return the predicted vector computed
 as the arithmetic mean of all elements belonging to a same motif.
}
\details{
Prediction is computed using arithmetic mean \code{amean} by motif
 \code{bymot} in a whole (WITHOUT taking into account species contribution).
  The elements belonging to a same motif are divided into \code{jack[2]}
   subsets of \code{jack[1]} elements. Prediction is computed by excluding
    \code{jack[1]} elements, of which the element to predict. If the total
     number of elements belonging to the motif is lower than
      \code{jack[1]*jack[2]}, prediction is computed by Leave-One-Out (LOO).
}
\keyword{internal}
