% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.ext.R
\name{predict_twin_fct}
\alias{predict_twin_fct}
\title{Compute assembly functioning and associated statistiques
         using the Clustering model
         for Interaction effet (alpha), Composition effect (beta),
           and assembly Functioning as Alpha * Beta * Fscale}
\usage{
predict_twin_fct(tree.cal, mOccur, alpha, beta,
                        xpr       = rep(1, length(alpha)),
                        fscale    = 1,
                        titre     = "",
                        opt.alpha = "gmean",
                        opt.beta  = "amean",
                        opt.mod   = "byelt",
                        opt.jack  = FALSE,    jack = c(2,5) )
}
\arguments{
\item{tree.cal}{gghhh}

\item{mOccur}{gghhh}

\item{alpha}{gghhh}

\item{beta}{gghhh}

\item{xpr}{gghhh}

\item{fscale}{gghhh}

\item{titre}{gghhh}

\item{opt.alpha}{gghhh}

\item{opt.beta}{gghhh}

\item{opt.mod}{gghhh}

\item{opt.jack}{gghhh}

\item{jack}{gghhh}
}
\value{
gghhh
}
\description{
Take a numeric f.
}
\details{
dd
}
