% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{pvalue.dependent.R2mse}
\alias{pvalue.dependent.R2mse}
\title{Test for the dependence of two R2}
\usage{
pvalue.dependent.R2mse(mprd, obs)
}
\arguments{
\item{mprd}{v2 v3 reg A linear model obtained using `lm` function.}

\item{obs}{v2 v3 reg A linear model obtained using `lm` function.}
}
\value{
Return a `list` containing a vector `x` of x-values (ranging from `xlim[1]` to `xlim[2]`), the corresponding values `y.confidence` for confidence interval, the corresponding values `y.prediction` for prediction interval.
}
\description{
`pvalue.dependent.R2mse` tests for the dependence of two R2.
}
\details{
None.
}
