% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{pvalue_dependent_R2mse}
\alias{pvalue_dependent_R2mse}
\title{Test for the dependence of two R2}
\usage{
pvalue_dependent_R2mse(mprd, obs)
}
\arguments{
\item{mprd}{a matrix of numeric vectors.}

\item{obs}{the reference vector of length equals to \code{dim(mprd)[2]}.}
}
\value{
Return a vector of p-value of length equals
 to \code{dim(mprd)[1]}.
}
\description{
\code{pvalue_dependent_R2mse} tests for
the dependence of two R2.
}
\details{
The function compare R2 obtained by the models
 \code{mprd[i, ] ~ obs}.
}
\keyword{internal}
