% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separating.ext.R
\name{rm_dual_assemblies}
\alias{rm_dual_assemblies}
\title{Remove species assemblies that were in several exemplar}
\usage{
rm_dual_assemblies(mat, fct,
                          xpr = rep(1, length(fct)),
                          opt.mean = "amean")
}
\arguments{
\item{mat}{bb}

\item{fct}{vv}

\item{xpr}{vv}

\item{opt.mean}{vv}
}
\value{
ff
}
\description{
cc
}
