% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separating.ext.R
\name{rm_elements}
\alias{rm_elements}
\title{Remove one or more species from data}
\usage{
rm_elements(mOccur, fct, elements)
}
\arguments{
\item{mOccur}{dd}

\item{fct}{dd}

\item{elements}{dd}
}
\value{
dd
}
\description{
ddd
}
