% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error RMSE) of two numeric vectors}
\usage{
rmse(x, y)
}
\arguments{
\item{x, y}{numeric vectors of same length.}
}
\value{
Return the Root Mean Square Error (RMSE).
}
\description{
Take two numeric vectors and return their
Root Mean Square Error (RMSE).
}
\details{
None.
}
\keyword{internal}
