% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{rss}
\alias{rss}
\title{Compute the Residual Sum of Square (RSS) of two numeric vectors}
\usage{
rss(x, y)
}
\arguments{
\item{x}{A logical/numeric vector.}

\item{y}{A logical/numeric vector.}
}
\value{
Return the Residual Sum of Square (RSS).
}
\description{
`rss` returns the Residual Sum of Square (RSS) of two numeric vectors.
}
\details{
None.
}
