% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.ext.R
\name{rss_cluster_elt}
\alias{rss_cluster_elt}
\title{ddd}
\usage{
rss_cluster_elt(affectElt, mOccur, fct, xpr,
                       opt.mean, opt.mod)
}
\arguments{
\item{affectElt}{ff}

\item{mOccur}{ff}

\item{fct}{ff}

\item{xpr}{ff}

\item{opt.mean}{ff}

\item{opt.mod}{gg}
}
\value{
gg
}
\description{
lll
}
\keyword{internal}
