% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.int.R
\name{shift_affectElt}
\alias{shift_affectElt}
\title{Indexes the species clusters by decreasing structuring effect}
\usage{
shift_affectElt(affectElt)
}
\arguments{
\item{affectElt}{gg}
}
\value{
ff
}
\description{
ddd
}
\details{
dd
}
\keyword{internal}
