% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.ext.R
\name{significant_assemblies}
\alias{significant_assemblies}
\title{dd}
\usage{
significant_assemblies(tree.ref, mOccur, fct, nbOpt,
xpr      = rep(1, length(fct)),
opt.ind  = "Jaccard",
opt.meth = "divisive",
opt.mean = "amean",
opt.mod  = "byelt",
opt.plot = TRUE)
}
\arguments{
\item{tree.ref}{gghhh}

\item{mOccur}{gghhh}

\item{fct}{gghhh}

\item{nbOpt}{gghhh}

\item{xpr}{gghhh}

\item{opt.ind}{gghhh}

\item{opt.meth}{gghhh}

\item{opt.mean}{gghhh}

\item{opt.mod}{gghhh}

\item{opt.plot}{gghhh}
}
\value{
gg
}
\description{
dd
}
