% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.ext.R
\name{significant_elements}
\alias{significant_elements}
\title{mmm}
\usage{
significant_elements(tree.ref, mOccur, fct, nbOpt,
                            xpr      = rep(1, length(fct)),
                            opt.ind  = "Jaccard",
                            opt.meth = "divisive",
                            opt.mean = "amean",
                            opt.mod  = "byelt",
                            opt.plot = TRUE)
}
\arguments{
\item{tree.ref}{ff}

\item{mOccur}{gg}

\item{fct}{gg}

\item{nbOpt}{gg}

\item{xpr}{gg}

\item{opt.ind}{gg}

\item{opt.meth}{gg}

\item{opt.mean}{gg}

\item{opt.mod}{gg}

\item{opt.plot}{gg}
}
\value{
gg
}
\description{
fff
}
\details{
"Jaccard",
opt.ind can be : "Czekanowski_Dice", "Folkes_Mallows", "Hubert", "Jaccard",
"Kulczynski", "McNemar", "Phi", "Precision", "Rand", "Recall",
"Rogers_Tanimoto", "Russel_Rao", "Sokal_Sneath1", "Sokal_Sneath2"
}
