% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.int.R
\name{sort_tree}
\alias{sort_tree}
\title{Sort the resulting file of tree}
\usage{
sort_tree(tree, index.return = FALSE)
}
\arguments{
\item{tree}{ff}

\item{index.return}{ff}
}
\value{
ff
}
\description{
ddd
}
\details{
dd
}
\keyword{internal}
