% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{test.dependent.R2}
\alias{test.dependent.R2}
\title{Test for the dependence of two R2}
\usage{
test.dependent.R2(v1, v2, v3, n = length(v1))
}
\arguments{
\item{v1, v2, v3}{three numeric vectors of same length n, or three coefficients of determination. If inputs are three coefficients of determination, length of vector n must be specified.}

\item{n}{the length of vectors. n must be specified if the inputs are three coefficients of determination.}
}
\value{
Return a p-value.
}
\description{
Tests for the dependence of two R2.
}
\details{
None.
}
