% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{test.dependent.R2mse}
\alias{test.dependent.R2mse}
\title{Test for the dependence of two R2}
\usage{
test.dependent.R2mse(prd2, prd1, obs)
}
\arguments{
\item{prd2, prd1}{three numeric vectors of same length (opt = "x"), or three coefficients of determination if opt = "R2".}

\item{obs}{an option for indicating if the inputs are vectors (opt = "x") or coefficients of determination (opt = "R2")}
}
\value{
Return a p-value.
}
\description{
Tests for the dependence of two R2.
}
\details{
None.
}
