% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{test.posthoc}
\alias{test.posthoc}
\title{Test posthoc of variance analysis}
\usage{
test.posthoc(x, clusters, pvalue = 0.05)
}
\arguments{
\item{x}{a numeric vector.}

\item{clusters}{a vector of factors of same length as `x`.}

\item{pvalue}{a marginal p-value.}
}
\value{
Return a `data.frame` containing the name of the groups `motif`, the group size `number`, the group mean `mean` and standard deviation `sd`, and the two-to-two differences expressed in the form of letters `group`.
}
\description{
Proceed to a variance analysis of the vector of data `x` associated with the vector of factors `clusters`, and return a `data.frame` containing the name of the groups `motif`, the group size `number`, the group mean `mean` and standard deviation `sd`, and the two-to-two differences expressed in the form of letters `group`.
}
\details{
Posthoc test uses Tukey method. Different groups are sorted by decreasing means. Letter rank increases with decreasing means.
}
