% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{test_dependent_R2mse}
\alias{test_dependent_R2mse}
\title{Test for the dependence of two R2}
\usage{
test_dependent_R2mse(prd2, prd1, obs)
}
\arguments{
\item{prd2, prd1, obs}{three numeric vectors of same length.}
}
\value{
Return a p-value.
}
\description{
Tests for the dependence of two R2.
}
\details{
Be careful, the three vectors are not symmetrical.
The function compare R2 obtained by the models \code{prd1 ~ obs}
and \code{prd2 ~ obs}.
}
\keyword{internal}
