% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{wamean}
\alias{wamean}
\title{Weighted arithmetical mean}
\usage{
wamean(x, w)
}
\arguments{
\item{x}{A logical/numeric vector.}

\item{w}{A vector of numeric weights, of same length as x.}
}
\value{
Return the weighted arithmetic mean of a vector.
}
\description{
`wamean` takes a logical/numeric vector and returns the weighted arithmetic mean.
}
\details{
None.
}
