% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{wgmean}
\alias{wgmean}
\title{Weighted geometric mean}
\usage{
wgmean(x, w)
}
\arguments{
\item{x}{A logical/numeric vector.}

\item{w}{A vector of weights, of same length as x.}
}
\value{
Return the weighted geometric standard deviation of a vector.
}
\description{
`wgmean` takes a logical/numeric vector and returns its weighted geometric mean.
}
\details{
None.
}
