% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{AIC}
\alias{AIC}
\title{AIC of two numeric vectors}
\usage{
AIC(x, y, nbK)
}
\arguments{
\item{x, y}{numeric vectors of same length.}

\item{nbK}{an integer. \code{nbK}should be higher than 1: return
an error if not.}
}
\value{
Return the the Akaike Information Criterion (AIC).
}
\description{
Take two numeric vectors and return their Akaike Information
 Criterion (AIC) computed from the Residuals Sum of Square and
 the number of used parameters.
}
\details{
None.
}
\keyword{internal}
