% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{AICc}
\alias{AICc}
\title{Compute the AICc from the RSS value.}
\usage{
AICc(x, y, nbK)
}
\arguments{
\item{x}{A logical/numeric vector.}

\item{y}{A logical/numeric vector.}

\item{nbK}{A scalar. `nbK`should be higher than `length(x)`. Returns an error if not.}
}
\value{
Return the the Akaike Information Criterion (AIC).
}
\description{
`AICc` returns the Akaike Information Criterion (AIC) computed from the Residuals Sum of Square and the number of used parameters, and corrected for .
}
\details{
Be careful, the `pmse` function is not symmetrical. The first argument is vector of estimated, modelled or predicted values, the second argument is the vector of reference.
}
