% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{amean}
\alias{amean}
\title{Arithmetic mean}
\usage{
amean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A logical/numeric vector.}

\item{na.rm}{The option.}
}
\value{
Return the arithmetic mean of a vector, as standard 'mean' function.
}
\description{
`amean` takes a logical/numeric vector and returns the arithmetic mean.
}
\details{
None.
}
\examples{
x <- c(1.2, 2.2, 2.1, NA, 2.6)
amean(x, na.rm = FALSE)
amean(x, na.rm = TRUE)
amean(x)

}
