% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{argmax}
\alias{argmax}
\title{Returns the index of the maximum values of a vector.}
\usage{
argmax(x)
}
\arguments{
\item{x}{A logical/numeric vector.}
}
\value{
Return the index of the maximum values of a vector.
}
\description{
`argmax` returns the index of the maximum values of a vector.
}
\details{
`argmax` function is as the standard `which.max` function.
}
