% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{argmin}
\alias{argmin}
\title{Index of the minimum values of a vector}
\usage{
argmin(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the index of the minimum values of a vector.
}
\description{
Take a numeric vector and return the index of its minimum values.
}
\details{
`argmin` function is as the standard `which.min` function.
}
