% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{asd}
\alias{asd}
\title{Arithmetical standard deviation}
\usage{
asd(x)
}
\arguments{
\item{x}{A logical/numeric vector.}
}
\value{
Return the arithmetic standard deviation of a vector, as standard `stats::sd` function.
}
\description{
`asd` takes a logical/numeric vector and returns the arithmetic standard deviation.
}
\details{
`asd` function is as standard `sd` function, but with option `na.rm = TRUE` by default.
}
