% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{asd}
\alias{asd}
\title{Arithmetical standard deviation}
\usage{
asd(x, na.rm = TRUE)
}
\arguments{
\item{x}{A logical/numeric vector.}

\item{na.rm}{The option `na.rm` indicates if `NA` values must be ignored (`na.rm = TRUE` by default) or not (`na.rm = FALSE`). Return `NA` if there are `NA` values and `na.rm = FALSE`.}
}
\value{
Return the arithmetic standard deviation of a vector, as standard `stats::sd` function.
}
\description{
`asd` takes a logical/numeric vector and returns the arithmetic standard deviation.
}
\details{
None.
}
