% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{first.argmin}
\alias{first.argmin}
\title{Index of the first minimum value of a vector}
\usage{
first.argmin(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the index of the first (the lowest index) minimum values of a vector.
}
\description{
Take a numeric vector and return the index of the first (the lowest index) minimum values.
}
\details{
`first.argmin` function is as the standard `min(which.min)` function. #'
}
