% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{gmean}
\alias{gmean}
\title{Geometric mean}
\usage{
gmean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A logical/numeric vector.}

\item{na.rm}{The option `na.rm` indicates if `NA` values must be ignored (`na.rm = TRUE` by default) or not (`na.rm = FALSE`). Return `NA` if there are `NA` values and `na.rm = FALSE`.}
}
\value{
Return the geometric mean of a vector.
}
\description{
`gmean` takes a logical/numeric vector and returns its geometric mean.
}
\details{
None.
}
