% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{gmean}
\alias{gmean}
\title{Geometric mean}
\usage{
gmean(x)
}
\arguments{
\item{x}{A logical/numeric vector.}
}
\value{
Return the geometric mean of a vector.
}
\description{
`gmean` takes a logical/numeric vector and returns its geometric mean.
}
\details{
None.
}
