% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{gsd}
\alias{gsd}
\title{Geometric standard deviation}
\usage{
gsd(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the geometric standard deviation of the vector \code{x}.
}
\description{
Take a numeric vector and return its geometric
standard deviation.
}
\details{
None.
}
\keyword{internal}
