% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{gsd}
\alias{gsd}
\title{Geometric standard deviation}
\usage{
gsd(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the geometric standard deviation of a vector.
}
\description{
Take a numeric vector and return its geometricstandard deviation.
}
\details{
None.
}
