% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{mean_fct}
\alias{mean_fct}
\title{Switch for arithmetic or geometric mean according to option "opt.mean"}
\usage{
mean_fct(x, opt.mean)
}
\arguments{
\item{x}{A logical/numeric vector.}

\item{opt.mean}{The option `opt.mean` equals `amean` or `gmean`. It has no default value.}
}
\value{
Return the arithmetic or geometric mean of a vector.
}
\description{
`mean_fct` takes a logical/numeric vector and returns its weighted geometric standard deviation.
}
\details{
None.
}
