% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{mse}
\alias{mse}
\title{Mean Square Error (MSE) of two numeric vectors}
\usage{
mse(x, y)
}
\arguments{
\item{x, y}{numeric vectors of same length.}
}
\value{
Return the Mean Square Error (MSE).
}
\description{
Take two numeric vectors and return their
Mean Square Error (MSE).
}
\details{
None.
}
\keyword{internal}
