% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{pmse}
\alias{pmse}
\title{Compute the Probability associated with the Coeficient of Determination (R2) of two numeric vectors}
\usage{
pmse(prd, obs, nbK)
}
\arguments{
\item{prd}{A logical/numeric vector.}

\item{obs}{A logical/numeric vector.}

\item{nbK}{A scalar. `nbK`should be higher than 1. Returns an error if not.}
}
\value{
Return the Probability associated with the Coeficient of Determination (R2).
}
\description{
`pmse` returns the Probability associated with the Coefficient of Determination (R2) of two numeric vectors.
}
\details{
Be careful, the `pmse` function is not symmetrical. The first argument `prd` is vector of estimated, modelled or predicted values, the second argument `obs` is the vector of reference.
}
