% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{rmse}
\alias{rmse}
\title{Compute the Root Mean Square Error RMSE) of two numeric vectors}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{A logical/numeric vector.}

\item{y}{A logical/numeric vector.}
}
\value{
Return the Root Mean Square Error (RMSE).
}
\description{
`rmse` returns the Root Mean Square Error (RMSE) of two numeric vectors.
}
\details{
None.
}
