% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{rss}
\alias{rss}
\title{Residual Sum of Square (RSS) of two numeric vectors}
\usage{
rss(x, y)
}
\arguments{
\item{x, y}{numeric vectors of same length.}
}
\value{
Return the Residual Sum of Square (RSS).
}
\description{
Take two numeric vectors and return their
Residual Sum of Square (RSS).
}
\details{
None.
}
\keyword{internal}
