% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{stirling}
\alias{stirling}
\title{Number of Stirling of second kind}
\usage{
stirling(n)
}
\arguments{
\item{n}{a positive integer.}
}
\value{
Return a vector of integer,
 of which the \code{names} are the number \code{k} of clusters,
  and the \code{values} are the number of partitions
  into \code{k} clusters possibly combinated with \code{n} elements.
}
\description{
Take an integer and return the number
 of Stirling of second kind, that is the number
 of all possible partitions into k clusters among n elements
}
\details{
None.
}
\keyword{internal}
