% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{wamean}
\alias{wamean}
\title{Weighted arithmetical mean}
\usage{
wamean(x, w)
}
\arguments{
\item{x}{a numeric vector.}

\item{w}{a vector of numeric weights, of same length as x.}
}
\value{
Return the weighted arithmetic mean of a vector.
}
\description{
Take a numeric vector and return the weighted arithmetic mean.
}
\details{
None.
}
