% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.Stats.R
\name{wasd}
\alias{wasd}
\title{Weighted arithmetic standard deviation}
\usage{
wasd(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{A logical/numeric vector.}

\item{w}{A vector of numeric weights, of same length as x.}

\item{na.rm}{The option `na.rm` indicates if `NA` values must be ignored (`na.rm = TRUE` by default) or not (`na.rm = FALSE`). Return `NA` if there are `NA` values and `na.rm = FALSE`.}
}
\value{
Return the weighted arithmetic standard deviation of a vector.
}
\description{
`wasd` takes a logical/numeric vector and returns the weighted arithmetic standard deviation.
}
\details{
None.
}
