% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{wgmean}
\alias{wgmean}
\title{Weighted geometric mean}
\usage{
wgmean(x, w)
}
\arguments{
\item{x}{a numeric vector.}

\item{w}{a vector of weights, of same length as x.}
}
\value{
Return the weighted geometric standard deviation of the vector \code{x}.
}
\description{
Take a numeric vector and return its weighted geometric mean.
}
\details{
None.
}
\keyword{internal}
