% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.int.R
\name{wgsd}
\alias{wgsd}
\title{Weighted geometric standard deviation}
\usage{
wgsd(x, w)
}
\arguments{
\item{x}{a numeric vector.}

\item{w}{a vector of numeric weights, of same length as \code{x}.}
}
\value{
Return the weighted geometric standard deviation
of the vector \code{x}.
}
\description{
Take a numeric vector and return its weighted geometric
 standard deviation.
}
\details{
None.
}
\keyword{internal}
